refactor(surface-nix-host): extract spawn helpers to process.ts
`runProgress` and `runCapture` were private to nixCopy.ts. A second
consumer (the arch probe added in the next commit) needs them too —
extracting them out makes the duplication impossible. nixCopy.ts now
imports both.

Pure move; both helpers keep their signatures, their close-event-flush
semantics, and their stderr-fanout behaviour.
